refactor: Review 도메인 Package 구조 Query와 Command로 분리
