+++ src/main/java/project/reviewing/review/command/domain/Review.java
@@ -1,4 +1,4 @@
-package project.reviewing.review.domain;
+package project.reviewing.review.command.domain;
 
 import lombok.AccessLevel;
 import lombok.Getter;
