+++ src/main/java/project/reviewing/review/command/domain/ReviewRepository.java
@@ -1,4 +1,4 @@
-package project.reviewing.review.domain;
+package project.reviewing.review.command.domain;
 
 import org.springframework.data.repository.Repository;
