+++ src/test/java/project/reviewing/integration/review/application/ReviewServiceTest.java
@@ -12,10 +12,10 @@ import project.reviewing.member.command.domain.Job;
 import project.reviewing.member.command.domain.Member;
 import project.reviewing.member.command.domain.Reviewer;
 import project.reviewing.member.exception.MemberNotFoundException;
-import project.reviewing.review.application.ReviewService;
-import project.reviewing.review.application.response.SingleReviewReadResponse;
-import project.reviewing.review.domain.Review;
-import project.reviewing.review.domain.ReviewStatus;
+import project.reviewing.review.command.application.ReviewService;
+import project.reviewing.review.command.application.response.SingleReviewReadResponse;
+import project.reviewing.review.command.domain.Review;
+import project.reviewing.review.command.domain.ReviewStatus;
 import project.reviewing.review.exception.InvalidReviewException;
 import project.reviewing.review.exception.ReviewNotFoundException;
 import project.reviewing.review.presentation.request.ReviewCreateRequest;
