+++ src/test/java/project/reviewing/unit/ControllerTest.java
@@ -18,7 +18,7 @@ import project.reviewing.auth.presentation.RefreshInterceptor;
 import project.reviewing.member.command.application.MemberService;
 import project.reviewing.member.query.application.MemberQueryService;
 import project.reviewing.tag.query.application.TagQueryService;
-import project.reviewing.review.application.ReviewService;
+import project.reviewing.review.command.application.ReviewService;
 
 @WebMvcTest(includeFilters = @Filter(type = FilterType.ANNOTATION, classes = RestController.class))
 @Import({
